Restructured project for greater clarity. Changed path and file names, header guards.
